krackns_agent/requirements.txt
# Core bot and web
python-telegram-bot==13.15
streamlit
python-dotenv
requests

# APK analysis
apkutils
androguard
pyaxmlparser

# Static & dynamic analysis
jsbeautifier
yara-python
frida-tools
frida

# Optional: for type hints and linting
mypy

# Optional: for improved logging and debugging
loguru

# Optional: for YAML config support
pyyaml

# Optional: for database support
sqlalchemy

# For decompilation (JADX CLI must be installed separately)
jadx